trfsintptr: import from nix (accepted)
Supress a warning mostly.

Reference: /n/patches.lsub.org/patch/trfsintptr
Reviewed-on: http://nxm.coreboot.org/review/259
Tested-by: Jenkins QA

+++ sys/src/cmd/trfs.c
@@ -41,7 +41,7 @@ static ulong
 getaux(Fid* fp)
 {
 	assert(sizeof(ulong) <= sizeof(fp->aux));
-	return (ulong)fp->aux;
+	return (uintptr)fp->aux;
 }
 
 static void
